Refactor Visitor and Transformer to include an instance graph
