+++ packages/type-traverser/src/transformer/Transformers.ts
@@ -12,6 +12,8 @@ import type {
   UnionType,
 } from "..";
 import type { InterfaceInstanceNode } from "../instanceGraph/nodes/InterfaceInstanceNode";
+import type { PrimitiveInstanceNode } from "../instanceGraph/nodes/PrimitiveInstanceNode";
+import type { UnionInstanceNode } from "../instanceGraph/nodes/UnionInstanceNode";
 
 export type GetTransformedChildrenFunction<TransformedChildrenType> =
   () => Promise<TransformedChildrenType>;
@@ -22,7 +24,8 @@ export type SetReturnPointerFunction<ReturnType> = (
 
 export type InterfaceTransformerFunction<
   Types extends TraverserTypes<any>,
-  Type extends InterfaceType<keyof Types>,
+  TypeName extends keyof Types,
+  Type extends InterfaceType<keyof Types> & Types[TypeName],
   ReturnType extends InterfaceReturnType<Type>,
   Context,
 > = (
@@ -31,13 +34,14 @@ export type InterfaceTransformerFunction<
     [PropertyName in keyof ReturnType["properties"]]: ReturnType["properties"][PropertyName];
   }>,
   setReturnPointer: SetReturnPointerFunction<ReturnType["return"]>,
-  node: InterfaceInstanceNode<Types, Type>,
+  node: InterfaceInstanceNode<Types, TypeName, Type>,
   context: Context,
 ) => Promise<ReturnType["return"]>;
 
 export type InterfaceTransformerPropertyFunction<
   Types extends TraverserTypes<any>,
-  Type extends InterfaceType<keyof Types>,
+  TypeName extends keyof Types,
+  Type extends InterfaceType<keyof Types> & Types[TypeName],
   ReturnTypes extends TransformerReturnTypes<Types>,
   ReturnType extends InterfaceReturnType<Type>,
   PropertyName extends keyof Type["properties"],
@@ -47,20 +51,29 @@ export type InterfaceTransformerPropertyFunction<
   getTransfromedChildren: GetTransformedChildrenFunction<
     ReturnTypes[Type["properties"][PropertyName]]["return"]
   >,
+  node: InterfaceInstanceNode<Types, TypeName, Type>,
   context: Context,
 ) => Promise<ReturnType["properties"][PropertyName]>;
 
 export type InterfaceTransformerDefinition<
   Types extends TraverserTypes<any>,
-  Type extends InterfaceType<keyof Types>,
+  TypeName extends keyof Types,
+  Type extends InterfaceType<keyof Types> & Types[TypeName],
   ReturnTypes extends TransformerReturnTypes<Types>,
   ReturnType extends InterfaceReturnType<Type>,
   Context,
 > = {
-  transformer: InterfaceTransformerFunction<Types, Type, ReturnType, Context>;
+  transformer: InterfaceTransformerFunction<
+    Types,
+    TypeName,
+    Type,
+    ReturnType,
+    Context
+  >;
   properties: {
     [PropertyName in keyof Type["properties"]]: InterfaceTransformerPropertyFunction<
       Types,
+      TypeName,
       Type,
       ReturnTypes,
       ReturnType,
@@ -72,7 +85,8 @@ export type InterfaceTransformerDefinition<
 
 export type UnionTransformerFunction<
   Types extends TraverserTypes<any>,
-  Type extends UnionType<keyof Types>,
+  TypeName extends keyof Types,
+  Type extends UnionType<keyof Types> & Types[TypeName],
   ReturnTypes extends TransformerReturnTypes<Types>,
   ReturnType extends UnionReturnType,
   Context,
@@ -82,31 +96,45 @@ export type UnionTransformerFunction<
     ReturnTypes[Type["typeNames"]]["return"]
   >,
   setReturnPointer: SetReturnPointerFunction<ReturnType["return"]>,
+  node: UnionInstanceNode<Types, TypeName, Type>,
   context: Context,
 ) => Promise<ReturnType["return"]>;
 
 export type UnionTransformerDefinition<
   Types extends TraverserTypes<any>,
-  Type extends UnionType<keyof Types>,
+  TypeName extends keyof Types,
+  Type extends UnionType<keyof Types> & Types[TypeName],
   ReturnTypes extends TransformerReturnTypes<Types>,
   ReturnType extends UnionReturnType,
   Context,
-> = UnionTransformerFunction<Types, Type, ReturnTypes, ReturnType, Context>;
+> = UnionTransformerFunction<
+  Types,
+  TypeName,
+  Type,
+  ReturnTypes,
+  ReturnType,
+  Context
+>;
 
 export type PrimitiveTransformerFunction<
-  Type extends PrimitiveType,
+  Types extends TraverserTypes<any>,
+  TypeName extends keyof Types,
+  Type extends PrimitiveType & Types[TypeName],
   ReturnType extends PrimitiveReturnType,
   Context,
 > = (
   originalData: Type["type"],
+  node: PrimitiveInstanceNode<Types, TypeName, Type>,
   context: Context,
 ) => Promise<ReturnType["return"]>;
 
 export type PrimitiveTransformerDefinition<
-  Type extends PrimitiveType,
+  Types extends TraverserTypes<any>,
+  TypeName extends keyof Types,
+  Type extends PrimitiveType & Types[TypeName],
   ReturnType extends PrimitiveReturnType,
   Context,
-> = PrimitiveTransformerFunction<Type, ReturnType, Context>;
+> = PrimitiveTransformerFunction<Types, TypeName, Type, ReturnType, Context>;
 
 export type TransformerDefinition<
   Types extends TraverserTypes<any>,
@@ -117,6 +145,7 @@ export type TransformerDefinition<
   ? ReturnTypes[TypeName] extends InterfaceReturnType<Types[TypeName]>
     ? InterfaceTransformerDefinition<
         Types,
+        TypeName,
         Types[TypeName],
         ReturnTypes,
         ReturnTypes[TypeName],
@@ -127,6 +156,7 @@ export type TransformerDefinition<
   ? ReturnTypes[TypeName] extends UnionReturnType
     ? UnionTransformerDefinition<
         Types,
+        TypeName,
         Types[TypeName],
         ReturnTypes,
         ReturnTypes[TypeName],
@@ -136,6 +166,8 @@ export type TransformerDefinition<
   : Types[TypeName] extends PrimitiveType
   ? ReturnTypes[TypeName] extends PrimitiveReturnType
     ? PrimitiveTransformerDefinition<
+        Types,
+        TypeName,
         Types[TypeName],
         ReturnTypes[TypeName],
         Context
@@ -161,15 +193,23 @@ export type Transformers<
  */
 export type InterfaceTransformerInputDefinition<
   Types extends TraverserTypes<any>,
-  Type extends InterfaceType<keyof Types>,
+  TypeName extends keyof Types,
+  Type extends InterfaceType<keyof Types> & Types[TypeName],
   ReturnTypes extends TransformerReturnTypes<Types>,
   ReturnType extends InterfaceReturnType<Type>,
   Context,
 > = {
-  transformer: InterfaceTransformerFunction<Types, Type, ReturnType, Context>;
+  transformer: InterfaceTransformerFunction<
+    Types,
+    TypeName,
+    Type,
+    ReturnType,
+    Context
+  >;
   properties?: Partial<{
     [PropertyName in keyof Type["properties"]]: InterfaceTransformerPropertyFunction<
       Types,
+      TypeName,
       Type,
       ReturnTypes,
       ReturnType,
@@ -181,17 +221,27 @@ export type InterfaceTransformerInputDefinition<
 
 export type UnionTransformerInputDefinition<
   Types extends TraverserTypes<any>,
-  Type extends UnionType<keyof Types>,
+  TypeName extends keyof Types,
+  Type extends UnionType<keyof Types> & Types[TypeName],
   ReturnTypes extends TransformerReturnTypes<Types>,
   ReturnType extends UnionReturnType,
   Context,
-> = UnionTransformerFunction<Types, Type, ReturnTypes, ReturnType, Context>;
+> = UnionTransformerFunction<
+  Types,
+  TypeName,
+  Type,
+  ReturnTypes,
+  ReturnType,
+  Context
+>;
 
 export type PrimitiveTransformerInputDefinition<
-  Type extends PrimitiveType,
+  Types extends TraverserTypes<any>,
+  TypeName extends keyof Types,
+  Type extends PrimitiveType & Types[TypeName],
   ReturnType extends PrimitiveReturnType,
   Context,
-> = PrimitiveTransformerFunction<Type, ReturnType, Context>;
+> = PrimitiveTransformerFunction<Types, TypeName, Type, ReturnType, Context>;
 
 export type TransformerInputDefinition<
   Types extends TraverserTypes<any>,
@@ -202,6 +252,7 @@ export type TransformerInputDefinition<
   ? ReturnTypes[TypeName] extends InterfaceReturnType<Types[TypeName]>
     ? InterfaceTransformerInputDefinition<
         Types,
+        TypeName,
         Types[TypeName],
         ReturnTypes,
         ReturnTypes[TypeName],
@@ -212,6 +263,7 @@ export type TransformerInputDefinition<
   ? ReturnTypes[TypeName] extends UnionReturnType
     ? UnionTransformerInputDefinition<
         Types,
+        TypeName,
         Types[TypeName],
         ReturnTypes,
         ReturnTypes[TypeName],
@@ -221,6 +273,8 @@ export type TransformerInputDefinition<
   : Types[TypeName] extends PrimitiveType
   ? ReturnTypes[TypeName] extends PrimitiveReturnType
     ? PrimitiveTransformerInputDefinition<
+        Types,
+        TypeName,
         Types[TypeName],
         ReturnTypes[TypeName],
         Context
