+++ packages/type-traverser/src/transformer/transformerSubTraversers/TransformerUnionSubTraverser.ts
@@ -1,12 +1,13 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import type { TraverserTypes } from "..";
+import type { TraverserTypes } from "../../";
+import type { UnionInstanceNode } from "../../instanceGraph/nodes/UnionInstanceNode";
 import type {
   TransformerReturnTypes,
   UnionReturnType,
 } from "../TransformerReturnTypes";
 import type { UnionTransformerDefinition } from "../Transformers";
-import type { UnionTraverserDefinition } from "../traverser/TraverserDefinition";
-import type { UnionType } from "../traverser/TraverserTypes";
+import type { UnionTraverserDefinition } from "../../traverser/TraverserDefinition";
+import type { UnionType } from "../../traverser/TraverserTypes";
 import { transformerParentSubTraverser } from "./TransformerParentSubTraverser";
 import type { TransformerSubTraverserGlobals } from "./util/transformerSubTraverserTypes";
 
@@ -14,7 +15,7 @@ export async function transformerUnionSubTraverser<
   Types extends TraverserTypes<any>,
   TypeName extends keyof Types,
   ReturnTypes extends TransformerReturnTypes<Types>,
-  Type extends UnionType<keyof Types>,
+  Type extends UnionType<keyof Types> & Types[TypeName],
   ReturnType extends UnionReturnType,
   Context,
 >(
@@ -39,6 +40,7 @@ export async function transformerUnionSubTraverser<
         itemTypeName
       ] as unknown as UnionTransformerDefinition<
         Types,
+        TypeName,
         Type,
         ReturnTypes,
         ReturnType,
@@ -66,6 +68,10 @@ export async function transformerUnionSubTraverser<
         (input) => {
           resolve(input);
         },
+        globals.instanceGraph.getNodeFor(
+          item,
+          itemTypeName,
+        ) as unknown as UnionInstanceNode<Types, TypeName, Type>,
         globals.context,
       );
       resolve(transformedObject);
