+++ packages/type-traverser/src/transformer/transformerSubTraversers/util/timeout.ts
