+++ packages/type-traverser/src/visitor/visitorSubTraversers/VisitorInterfaceSubTraverser.ts
@@ -1,14 +1,15 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import type { InterfaceVisitorDefinition, TraverserTypes } from "..";
-import type { InterfaceTraverserDefinition } from "../traverser/TraverserDefinition";
-import type { InterfaceType } from "../traverser/TraverserTypes";
+import type { InterfaceVisitorDefinition, TraverserTypes } from "../../";
+import type { InterfaceInstanceNode } from "../../instanceGraph/nodes/InterfaceInstanceNode";
+import type { InterfaceTraverserDefinition } from "../../traverser/TraverserDefinition";
+import type { InterfaceType } from "../../traverser/TraverserTypes";
 import type { VisitorSubTraverserGlobals } from "./util/visitorSubTraverserTypes";
 import { visitorParentSubTraverser } from "./VisitorParentSubTraverser";
 
 export async function visitorInterfaceSubTraverser<
   Types extends TraverserTypes<any>,
   TypeName extends keyof Types,
-  Type extends InterfaceType<keyof Types>,
+  Type extends InterfaceType<keyof Types> & Types[TypeName],
   Context,
 >(
   item: Type["type"],
@@ -22,16 +23,27 @@ export async function visitorInterfaceSubTraverser<
   ] as InterfaceTraverserDefinition<Type>;
   const visitor = visitors[
     itemTypeName
-  ] as unknown as InterfaceVisitorDefinition<Types, Type, Context>;
+  ] as unknown as InterfaceVisitorDefinition<Types, TypeName, Type, Context>;
 
   await Promise.all([
-    visitor.visitor(item, globals.context),
+    visitor.visitor(
+      item,
+      globals.instanceGraph.getNodeFor(
+        item,
+        itemTypeName,
+      ) as unknown as InterfaceInstanceNode<Types, TypeName, Type>,
+      globals.context,
+    ),
     Promise.all(
       Object.entries(definition.properties).map(async ([propertyName]) => {
         const originalObject = item[propertyName];
         const originalPropertyDefinition = definition.properties[propertyName];
         const propertyVisitorPromise = visitor.properties[propertyName](
           originalObject,
+          globals.instanceGraph.getNodeFor(
+            item,
+            itemTypeName,
+          ) as unknown as InterfaceInstanceNode<Types, TypeName, Type>,
           globals.context,
         );
         let propertyTraverserPromise: Promise<void | void[]>;
