+++ packages/type-traverser/test/integration/avatar/AvatarBrokenTransformer.ts
