+++ packages/type-traverser/test/integration/avatar/AvatarErroringTransformer.ts
