+++ packages/type-traverser/test/integration/avatar/AvatarTraverserTypes.ts
