+++ packages/type-traverser/test/integration/avatar/sampleData.ts
